Trait rumbas_support::input::Input

pub trait Input: Clone {
    type Normal;

    fn to_normal(&self) -> Self::Normal;
    fn find_missing(&self) -> InputCheckResult;
    fn from_normal(normal: Self::Normal) -> Self;
    fn insert_template_value(&mut self, key: &str, val: &Value);
    fn files_to_load(&self) -> Vec<FileToLoad>;
    fn insert_loaded_files(&mut self, files: &HashMap<FileToLoad, LoadedFile>);
    fn dependencies(&self) -> HashSet<PathBuf>;

    fn to_normal_safe(&self) -> Result<Self::Normal, InputCheckResult> { ... }
}

Required Associated Types

type Normal

Required Methods

fn to_normal(&self) -> Self::Normal

This method assumes that it is called by a function that is initially called from to_normal_safe

fn find_missing(&self) -> InputCheckResult

Check the optional data

fn from_normal(normal: Self::Normal) -> Self

fn insert_template_value(&mut self, key: &str, val: &Value)

fn files_to_load(&self) -> Vec<FileToLoad>

fn insert_loaded_files(&mut self, files: &HashMap<FileToLoad, LoadedFile>)

fn dependencies(&self) -> HashSet<PathBuf>

Provided Methods

fn to_normal_safe(&self) -> Result<Self::Normal, InputCheckResult>

Method that safely convets the input type to the normal type